rename transaction_pool to top_pool
