+++ enclave/src/top_pool/listener.rs
@@ -21,7 +21,7 @@ use log::{debug, trace};
 use sgx_tstd::{collections::HashMap, fmt::Debug, hash, string::String, vec::Vec};
 use sp_runtime::traits;
 
-use crate::transaction_pool::{
+use crate::top_pool::{
     pool::{BlockHash, ChainApi, ExtrinsicHash},
     watcher::Watcher,
 };
