+++ enclave/src/top_pool/watcher.rs
@@ -26,7 +26,7 @@ use sp_runtime::traits;
 
 use sgx_tstd::hash;
 
-//use crate::transaction_pool::primitives::TransactionStatus
+//use crate::top_pool::primitives::TransactionStatus
 use crate::rpc::worker_api_direct;
 use substratee_worker_primitives::TransactionStatus;
